requirements.txt
impacket
tabulate
colorama
python-magic
chardet
